refactor: change PropTable exports
